Fix sometime failing test
The test assumed that the shuffled order of the list would
be X because it assumed no other calls to random took place.
Maybe some recent changes added a call to `random`, but the
test fails on my ubuntu machine when `make eunit` is run.

This change calls `seed` in the test to ensure that "determinsitic random"
behaviour occurs for the test.